TrabajoFinalGrado/requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
tensorflow==2.12.0
joblib==1.2.0
opencv-python==4.7.0.72
Pillow==9.5.0
tqdm==4.65.0
requests==2.31.0
beautifulsoup4==4.12.2